<!--
SPDX-FileCopyrightText: Copyright (c) 2025-2026, NVIDIA CORPORATION &
AFFILIATES. All rights reserved.
SPDX-License-Identifier: Apache-2.0

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
-->

## :snowflake: Code freeze for `release/1.8` and `v1.8.0` release

### What does this mean?
Only critical/hotfix level issues should be merged into `release/1.8`
until release (merging of this PR).

### What is the purpose of this PR?
- Update documentation
- Allow testing for the new release
- Enable a means to merge `release/1.8` into `main` for the release

+++ packages/nvidia_nat_test/src/nat/test/plugin.py
@@ -13,6 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import logging
 import os
 import random
 import subprocess
@@ -31,9 +32,9 @@ if typing.TYPE_CHECKING:
     import galileo.log_streams
     import galileo.projects
     import langsmith.client
-
-if typing.TYPE_CHECKING:
     from docker.client import DockerClient
+
+logger = logging.getLogger(__name__)
 
 
 def pytest_addoption(parser: pytest.Parser):
@@ -335,7 +336,11 @@ def langsmith_project_name_fixture(langsmith_client: "langsmith.client.Client", 
     langsmith_client.create_project(project_name)
     yield project_name
 
-    langsmith_client.delete_project(project_name=project_name)
+    if os.environ.get("NAT_CI_KEEP_LANGSMITH_PROJECTS") != "1":
+        try:
+            langsmith_client.delete_project(project_name=project_name)
+        except Exception:
+            logger.exception("Failed to delete project %s", project_name)
 
 
 @pytest.fixture(name="galileo_api_key", scope='session')
@@ -377,38 +382,6 @@ def galileo_log_stream_fixture(galileo_project: "galileo.projects.Project") -> "
     """
     import galileo.log_streams
     return galileo.log_streams.create_log_stream(project_id=galileo_project.id, name="test")
-
-
-@pytest.fixture(name="catalyst_keys", scope='session')
-def catalyst_keys_fixture(fail_missing: bool):
-    """
-    Use for integration tests that require RagaAI Catalyst credentials.
-    """
-    yield require_env_variables(
-        varnames=["CATALYST_ACCESS_KEY", "CATALYST_SECRET_KEY"],
-        reason="Catalyst integration tests require the `CATALYST_ACCESS_KEY` and `CATALYST_SECRET_KEY` environment "
-        "variables to be defined.",
-        fail_missing=fail_missing)
-
-
-@pytest.fixture(name="catalyst_project_name")
-def catalyst_project_name_fixture(catalyst_keys) -> str:
-    return os.environ.get("NAT_CI_CATALYST_PROJECT_NAME", "nat-e2e")
-
-
-@pytest.fixture(name="catalyst_dataset_name")
-def catalyst_dataset_name_fixture(catalyst_project_name: str, project_name: str) -> str:
-    """
-    We can't create and delete projects, but we can create and delete datasets, so use a unique dataset name
-    """
-    dataset_name = project_name.replace('.', '-')
-    yield dataset_name
-
-    from ragaai_catalyst import Dataset
-    ds = Dataset(catalyst_project_name)
-    datasets = ds.list_datasets()
-    if datasets and dataset_name in datasets:
-        ds.delete_dataset(dataset_name)
 
 
 @pytest.fixture(name="require_docker", scope='session')
